refactor : GameController의 CheckInt를 InputView의 책임으로 봐도 된다.
