+++ src/main/kotlin/baseball/domain/Game.kt
@@ -12,5 +12,4 @@ class Game{
         val count = Judgement().correctCount(randomBaseballNumber,numList)
         return count
     }
-
 }
